+++ code/game/machinery/suit_storage_unit.dm
@@ -1,4 +1,4 @@
-// SUIT STORAGE UNIT /////////////////
+// SUIT STORAGE UNIT /////////////////close_machine(
 /obj/machinery/suit_storage_unit
 	name = "suit storage unit"
 	desc = "An industrial unit made to hold space suits. It comes with a built-in UV cauterization mechanism. A small warning label advises that organic matter should not be placed into the unit."
@@ -6,8 +6,6 @@
 	icon_state = "close"
 	anchored = 1
 	density = 1
-	obj_integrity = 250
-	max_integrity = 250
 
 	var/obj/item/clothing/suit/space/suit = null
 	var/obj/item/clothing/head/helmet/space/helmet = null
@@ -34,8 +32,9 @@
 	mask_type = /obj/item/clothing/mask/breath
 
 /obj/machinery/suit_storage_unit/captain
-	suit_type = /obj/item/clothing/suit/space/hardsuit/captain
-	mask_type = /obj/item/clothing/mask/gas/sechailer
+	suit_type = /obj/item/clothing/suit/space/captain
+	helmet_type = /obj/item/clothing/head/helmet/space/captain
+	mask_type = /obj/item/clothing/mask/gas
 	storage_type = /obj/item/weapon/tank/jetpack/oxygen/captain
 
 /obj/machinery/suit_storage_unit/engine
@@ -61,10 +60,6 @@
 	storage_type = /obj/item/weapon/watertank/atmos
 
 /obj/machinery/suit_storage_unit/mining
-	suit_type = /obj/item/clothing/suit/hooded/explorer
-	mask_type = /obj/item/clothing/mask/gas/explorer
-
-/obj/machinery/suit_storage_unit/mining/eva
 	suit_type = /obj/item/clothing/suit/space/hardsuit/mining
 	mask_type = /obj/item/clothing/mask/breath
 
@@ -114,71 +109,62 @@
 		storage = new storage_type(src)
 	update_icon()
 
-/obj/machinery/suit_storage_unit/Destroy()
-	if(suit)
-		qdel(suit)
-		suit = null
-	if(helmet)
-		qdel(helmet)
-		helmet = null
-	if(mask)
-		qdel(mask)
-		mask = null
-	if(storage)
-		qdel(storage)
-		storage = null
-	return ..()
-
 /obj/machinery/suit_storage_unit/update_icon()
-	cut_overlays()
+	overlays.Cut()
 
 	if(uv)
 		if(uv_super)
-			add_overlay("super")
+			overlays += "super"
 		else if(occupant)
-			add_overlay("uvhuman")
+			overlays += "uvhuman"
 		else
-			add_overlay("uv")
+			overlays += "uv"
 	else if(state_open)
 		if(stat & BROKEN)
-			add_overlay("broken")
+			overlays += "broken"
 		else
-			add_overlay("open")
+			overlays += "open"
 			if(suit)
-				add_overlay("suit")
+				overlays += "suit"
 			if(helmet)
-				add_overlay("helm")
+				overlays += "helm"
 			if(storage)
-				add_overlay("storage")
+				overlays += "storage"
 	else if(occupant)
-		add_overlay("human")
+		overlays += "human"
 
 /obj/machinery/suit_storage_unit/power_change()
 	..()
 	if(!is_operational() && state_open)
-		open_machine()
-		dump_contents()
+		open_machine(dump = TRUE)
 	update_icon()
 
-/obj/machinery/suit_storage_unit/proc/dump_contents()
-	dropContents()
-	helmet = null
-	suit = null
-	mask = null
-	storage = null
-	occupant = null
+/obj/machinery/suit_storage_unit/open_machine(dump = FALSE)
+	state_open = TRUE
+	if(dump)
+		dropContents()
+		helmet = null
+		suit = null
+		mask = null
+		storage = null
+		occupant = null
+	update_icon()
 
-/obj/machinery/suit_storage_unit/deconstruct(disassembled = TRUE)
-	if(!(flags & NODECONSTRUCT))
-		open_machine()
-		dump_contents()
-		new /obj/item/stack/sheet/metal (loc, 2)
-	qdel(src)
+/obj/machinery/suit_storage_unit/ex_act(severity, target)
+	switch(severity)
+		if(1)
+			if(prob(50))
+				open_machine(dump = TRUE)
+			qdel(src)
+		if(2)
+			if(prob(50))
+				open_machine(dump = TRUE)
+				qdel(src)
 
-/obj/machinery/suit_storage_unit/MouseDrop_T(atom/A, mob/user)
-	if(user.stat || user.lying || !Adjacent(user) || !Adjacent(A) || !isliving(A))
+/obj/machinery/suit_storage_unit/MouseDrop_T(mob/target, mob/user)
+	if(user.stat || user.lying || !Adjacent(user) || !Adjacent(target))
 		return
-	var/mob/living/target = A
+
 	if(!state_open)
 		user << "<span class='warning'>The unit's doors are shut!</span>"
 		return
@@ -190,17 +176,11 @@
 		return
 
 	if(target == user)
-		user.visible_message("<span class='warning'>[user] starts squeezing into [src]!</span>", "<span class='notice'>You start working your way into [src]...</span>")
+		visible_message("<span class='warning'>[user] squeezes into [src]!</span>", "<span class='notice'>You squeeze into [src].</span>")
 	else
-		target.visible_message("<span class='warning'>[user] starts shoving [target] into [src]!</span>", "<span class='userdanger'>[user] starts shoving you into [src]!</span>")
+		visible_message("<span class='warning'>[user] starts putting [target] into [src]!</span>", "<span class='userdanger'>[user] starts shoving you into [src]!</span>")
 
-	if(do_mob(user, target, 30))
-		if(occupant || helmet || suit || storage)
-			return
-		if(target == user)
-			user.visible_message("<span class='warning'>[user] slips into [src] and closes the door behind them!</span>", "<span class=notice'>You slip into [src]'s cramped space and shut its door.</span>")
-		else
-			target.visible_message("<span class='warning'>[user] pushes [target] into [src] and shuts its door!<span>", "<span class='userdanger'>[user] shoves you into [src] and shuts the door!</span>")
+	if(do_mob(user, target, 10))
 		close_machine(target)
 		add_fingerprint(user)
 
@@ -217,14 +197,13 @@
 				occupant.adjustFireLoss(rand(10, 16))
 			if(iscarbon(occupant))
 				occupant.emote("scream")
-		addtimer(CALLBACK(src, .proc/cook), 50)
+		addtimer(src, "cook", 50, FALSE)
 	else
 		uv_cycles = initial(uv_cycles)
 		uv = FALSE
 		locked = FALSE
 		if(uv_super)
-			visible_message("<span class='warning'>[src]'s door creaks open with a loud whining noise. A cloud of foul black smoke escapes from its chamber.</span>")
-			playsound(src, 'sound/machines/airlock_alien_prying.ogg', 50, 1)
+			visible_message("<span class='warning'>With a loud whining noise, [src]'s door grinds open. A foul cloud of smoke emanates from the chamber.</span>")
 			helmet = null
 			qdel(helmet)
 			suit = null
@@ -236,37 +215,30 @@
 			// The wires get damaged too.
 			wires.cut_all()
 		else
-			if(!occupant)
-				visible_message("<span class='notice'>[src]'s door slides open. The glowing yellow lights dim to a gentle green.</span>")
-			else
-				visible_message("<span class='warning'>[src]'s door slides open, barraging you with the nauseating smell of charred flesh.</span>")
-			playsound(src, 'sound/machines/AirlockClose.ogg', 25, 1)
-			for(var/obj/item/I in src) //Scorches away blood and forensic evidence, although the SSU itself is unaffected
+			visible_message("<span class='warning'>With a loud whining noise, [src]'s door grinds open. A light cloud of steam escapes from the chamber.</span>")
+			for(var/obj/item/I in src)
 				I.clean_blood()
-				I.fingerprints = list()
-		open_machine(FALSE)
-		if(occupant)
-			dump_contents()
+		open_machine(dump = !!occupant)
 
 /obj/machinery/suit_storage_unit/proc/shock(mob/user, prb)
 	if(!prob(prb))
 		var/datum/effect_system/spark_spread/s = new /datum/effect_system/spark_spread
 		s.set_up(5, 1, src)
 		s.start()
-		if(electrocute_mob(user, src, src, 1, TRUE))
+		if(electrocute_mob(user, src, src))
 			return 1
 
 /obj/machinery/suit_storage_unit/relaymove(mob/user)
-	container_resist(user)
+	container_resist()
 
-/obj/machinery/suit_storage_unit/container_resist(mob/living/user)
+/obj/machinery/suit_storage_unit/container_resist()
+	var/mob/living/user = usr
 	add_fingerprint(user)
 	if(locked)
 		visible_message("<span class='notice'>You see [user] kicking against the doors of [src]!</span>", "<span class='notice'>You start kicking against the doors...</span>")
-		addtimer(CALLBACK(src, .proc/resist_open, user), 300)
+		addtimer(src, "resist_open", 300, FALSE, user)
 	else
-		open_machine()
-		dump_contents()
+		open_machine(dump = TRUE)
 
 /obj/machinery/suit_storage_unit/proc/resist_open(mob/user)
 	if(!state_open && occupant && (user in src) && user.stat == 0) // Check they're still here.
@@ -296,7 +268,7 @@
 			if(!user.drop_item())
 				return
 			mask = I
-		else
+		else if(istype(I, /obj/item))
 			if(storage)
 				user << "<span class='warning'>The auxiliary storage compartment is full!</span>"
 				return
@@ -306,8 +278,6 @@
 
 		I.loc = src
 		visible_message("<span class='notice'>[user] inserts [I] into [src]</span>", "<span class='notice'>You load [I] into [src].</span>")
-		update_icon()
-		return
 
 	if(panel_open && is_wire_tool(I))
 		wires.interact(user)
@@ -315,10 +285,10 @@
 		if(default_deconstruction_screwdriver(user, "panel", "close", I))
 			return
 	if(default_pry_open(I))
-		dump_contents()
 		return
 
-	return ..()
+	update_icon()
+	return
 
 /obj/machinery/suit_storage_unit/ui_interact(mob/user, ui_key = "main", datum/tgui/ui = null, force_open = 0, \
 										datum/tgui/master_ui = null, datum/ui_state/state = notcontained_state)
@@ -354,9 +324,7 @@
 			if(state_open)
 				close_machine()
 			else
-				open_machine(0)
-				if(occupant)
-					dump_contents() // Dump out contents if someone is in there.
+				open_machine(!!occupant) // Dump out contents if someone is in there.
 			. = TRUE
 		if("lock")
 			locked = !locked
@@ -367,8 +335,6 @@
 			else if(!helmet && !mask && !suit && !storage && !occupant)
 				return
 			else
-				if(occupant)
-					occupant << "<span class='userdanger'>[src]'s confines grow warm, then hot, then scorching. You're being burned [!occupant.stat ? "alive" : "away"]!</span>"
 				cook()
 				. = TRUE
 		if("dispense")
